fastapi-backend/requirements.txt
# FastAPI core dependencies
fastapi==0.115.0
uvicorn[standard]==0.32.0
pydantic==2.9.2
pydantic-settings==2.6.0

# Database (PostgreSQL)
psycopg2-binary==2.9.10
sqlalchemy==2.0.36

# Logging
structlog==24.4.0

# Environment variables
python-dotenv==1.0.1

# CORS and middleware
python-multipart==0.0.12

# HTTP client (for external API calls if needed)
httpx==0.27.2

# Existing backend dependencies (reuse from main project)
# Note: Make sure these match the main requirements.txt
requests>=2.31.0

# Authentication and security
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
python-multipart==0.0.12

# Monitoring and observability
sentry-sdk[fastapi]==2.19.2
prometheus-client==0.21.0
prometheus-fastapi-instrumentator==7.0.0
